+++ packages/components/atoms/f-form-field/src/components/_tests/FormField.test.js
@@ -2,12 +2,17 @@ import { shallowMount, mount } from '@vue/test-utils';
 import FormField from '../FormField.vue';
 import FormDropdown from '../FormDropdown.vue';
 import {
-    DEFAULT_INPUT_TYPE, VALID_INPUT_TYPES, VALID_LABEL_STYLES
+    DEFAULT_INPUT_TYPE, VALID_ICON_INPUT_TYPES, VALID_INPUT_TYPES, VALID_LABEL_STYLES, VALID_TRAILING_ICON_INPUT_TYPES
 } from '../../constants';
 
 const $style = {
-    'c-formField-field--noFocus': 'c-formField-field--noFocus'
+    'c-formField-field--noFocus': 'c-formField-field--noFocus',
+    'c-formField-field--small': 'c-formField-field--small',
+    'c-formField-field--medium': 'c-formField-field--medium',
+    'c-formField-field--large': 'c-formField-field--large'
 };
+
+const slot = '<div>ICON</div>';
 
 describe('FormField', () => {
     it('should be defined', () => {
@@ -234,6 +239,32 @@ describe('FormField', () => {
                 });
             });
         });
+
+        describe('fieldSize ::', () => {
+            it.each([
+                ['c-formField-field--small', 'small'],
+                ['c-formField-field--medium', 'medium'],
+                ['c-formField-field--large', 'large']
+            ])('should apply %s class to input when set to %s', (className, fieldSize) => {
+                // Arrange
+                const propsData = {
+                    fieldSize
+                };
+
+                // Act
+                const wrapper = shallowMount(FormField, {
+                    propsData,
+                    mocks: {
+                        $style
+                    }
+                });
+
+                const formInput = wrapper.find('input');
+
+                // Assert
+                expect(formInput.attributes('class')).toContain(className);
+            });
+        });
     });
 
     describe('computed :: ', () => {
@@ -252,6 +283,110 @@ describe('FormField', () => {
 
                 // Assert
                 expect(formInput.attributes('class')).toContain('c-formField-field--noFocus');
+            });
+        });
+
+        describe('isDisabled :: ', () => {
+            it.each([
+                [true, 'disabled'],
+                [false, null]
+            ])('should return %s if disabled attribute is %s', (expected, disabled) => {
+                // Arrange
+                const propsData = {
+                    disabled
+                };
+
+                // Act
+                const wrapper = shallowMount(FormField, { propsData });
+
+                // Assert
+                expect(wrapper.vm.isDisabled).toEqual(expected);
+            });
+        });
+
+        describe('isValidIconField :: ', () => {
+            it.each([
+                [true, 'email'],
+                [true, 'password'],
+                [true, 'dropdown'],
+                [true, 'number'],
+                [true, 'tel'],
+                [false, 'radio'],
+                [false, 'checkbox'],
+                [false, null]
+            ])('should return %s if `inputType` is %s', (expected, inputType) => {
+                // Arrange
+                const propsData = {
+                    inputType
+                };
+
+                // Act
+                const wrapper = shallowMount(FormField, { propsData });
+
+                // Assert
+                expect(wrapper.vm.isValidIconField).toEqual(expected);
+            });
+        });
+
+        describe('hasLeadingIcon :: ', () => {
+            // Arrange
+            const propsData = {
+                inputType: 'text'
+            };
+
+            it('should return `true` when slots contain `icon-leading`', () => {
+                // Act
+                const wrapper = shallowMount(FormField, {
+                    propsData,
+                    slots: {
+                        'icon-leading': slot
+                    }
+                });
+
+                // Assert
+                expect(wrapper.vm.hasLeadingIcon).toEqual(true);
+            });
+
+            it('should return `false` when slots do not contain `icon-leading`', () => {
+                // Act
+                const wrapper = shallowMount(FormField, {
+                    propsData,
+                    slots: null
+                });
+
+                // Assert
+                expect(wrapper.vm.hasLeadingIcon).toEqual(false);
+            });
+        });
+
+        describe('hasTrailingIcon :: ', () => {
+            // Arrange
+            const propsData = {
+                inputType: 'text'
+            };
+
+            it('should return `true` when slots contain `icon-trailing`', () => {
+                // Act
+                const wrapper = shallowMount(FormField, {
+                    propsData,
+                    slots: {
+                        'icon-trailing': slot
+                    }
+                });
+
+                // Assert
+                expect(wrapper.vm.hasTrailingIcon).toEqual(true);
+            });
+
+            it('should return `false` when slots do not contain `icon-trailing`', () => {
+                // Act
+                const wrapper = shallowMount(FormField, {
+                    propsData,
+                    slots: null
+                });
+
+                // Assert
+                expect(wrapper.vm.hasTrailingIcon).toEqual(false);
             });
         });
     });
@@ -284,7 +419,7 @@ describe('FormField', () => {
                 expect(wrapper.attributes('data-test-id')).toBe(`formfield-${attrsData.attrs.name}`);
             });
 
-            it('should include attribute `name` in the generated input data-test-id when it is set', () => {
+            it('should include attribute `name` and an `input` suffix` in the generated input data-test-id when it is set', () => {
                 // Arrange
                 const attrsData = {
                     attrs: {
@@ -298,6 +433,24 @@ describe('FormField', () => {
 
                 // Assert
                 expect(formInput.attributes('data-test-id')).toBe(`formfield-${attrsData.attrs.name}-input`);
+            });
+
+            it('should include attribute `name` and a `textarea` suffix in the generated textarea data-test-id when it is set', () => {
+                // Arrange
+                const attrs = {
+                    name: 'email'
+                };
+
+                const propsData = {
+                    inputType: 'textarea'
+                };
+
+                // Act
+                const wrapper = mount(FormField, { attrs, propsData });
+                const formInput = wrapper.find('textarea');
+
+                // Assert
+                expect(formInput.attributes('data-test-id')).toBe(`formfield-${attrs.name}-textarea`);
             });
 
             it('when name is set, it should be included in the generated testId.input data-test-id', () => {
@@ -351,6 +504,105 @@ describe('FormField', () => {
 
                 // Assert
                 expect(wrapper.emitted().input).toHaveLength(1);
+            });
+        });
+    });
+
+    describe('methods ::', () => {
+        describe('validateProps ::', () => {
+            let spy;
+
+            beforeEach(() => {
+                spy = jest.spyOn(global.console, 'error').mockImplementation(() => { });
+            });
+
+            afterEach(() => {
+                spy.mockRestore();
+            });
+
+            describe('when `hasLeadingIcon` is true', () => {
+                it.each(VALID_ICON_INPUT_TYPES)('should not throw an error when `inputType` is set to %s', inputType => {
+                    // Arrange
+                    const propsData = { inputType };
+
+                    // Act & Assert
+                    expect(() => {
+                        shallowMount(FormField, {
+                            propsData,
+                            slots: {
+                                'icon-leading': slot
+                            }
+                        });
+                    }).not.toThrowError();
+                });
+
+                it.each(['textarea', 'checkbox', 'radio'])('should throw an error when `inputType` is set to %s', inputType => {
+                    // Arrange
+                    const propsData = { inputType };
+
+                    // Act & Assert
+                    expect(() => {
+                        shallowMount(FormField, {
+                            propsData,
+                            slots: {
+                                'icon-leading': slot
+                            }
+                        });
+                    }).toThrowError(`Form field is set to have inputType="${inputType}", but icons can only be displayed one of the following inputTypes: "${VALID_ICON_INPUT_TYPES.join('", "')}"`);
+                });
+            });
+
+            describe('when `hasTrailingIcon` is true', () => {
+                it.each([
+                    'text',
+                    'email',
+                    'password',
+                    'number',
+                    'tel'
+                ])('should not throw an error when `inputType` is set to %s', inputType => {
+                    // Arrange
+                    const propsData = { inputType };
+
+                    // Act & Assert
+                    expect(() => {
+                        shallowMount(FormField, {
+                            propsData,
+                            slots: {
+                                'icon-trailing': slot
+                            }
+                        });
+                    }).not.toThrowError();
+                });
+
+                it.each(['textarea', 'checkbox', 'radio'])('should throw an error when `inputType` is set to %s', inputType => {
+                    // Arrange
+                    const propsData = { inputType };
+
+                    // Act & Assert
+                    expect(() => {
+                        shallowMount(FormField, {
+                            propsData,
+                            slots: {
+                                'icon-trailing': slot
+                            }
+                        });
+                    }).toThrowError(`Form field is set to have inputType="${inputType}", but icons can only be displayed one of the following inputTypes: "${VALID_ICON_INPUT_TYPES.join('", "')}"`);
+                });
+
+                it('should throw an error when `inputType` is set to `dropdown`', () => {
+                    // Arrange
+                    const propsData = { inputType: 'dropdown' };
+
+                    // Act & Assert
+                    expect(() => {
+                        shallowMount(FormField, {
+                            propsData,
+                            slots: {
+                                'icon-trailing': slot
+                            }
+                        });
+                    }).toThrowError(`Form field is set to have inputType="dropdown", but trailing icons can only be displayed one of the following inputTypes: "${VALID_TRAILING_ICON_INPUT_TYPES.join('", "')}"`);
+                });
             });
         });
     });