+++ packages/components/atoms/f-form-field/src/constants.js
@@ -1,7 +1,14 @@
 export const VALID_INPUT_TYPES = ['text', 'email', 'password', 'radio', 'checkbox', 'number', 'tel'];
 export const CUSTOM_INPUT_TYPES = ['dropdown', 'textarea'];
+
+export const VALID_TRAILING_ICON_INPUT_TYPES = ['text', 'email', 'password', 'number', 'tel'];
+export const VALID_ICON_INPUT_TYPES = VALID_TRAILING_ICON_INPUT_TYPES.concat('dropdown');
+
 export const DEFAULT_INPUT_TYPE = 'text';
 
 export const VALID_LABEL_STYLES = ['default', 'inline', 'inlineNarrow'];
+
+export const DEFAULT_FIELD_SIZE = 'medium';
+export const VALID_FIELD_SIZES = ['small', 'medium', 'large'];
 
 export const MOBILE_WIDTH = 768;