+++ packages/components/atoms/f-form-field/stories/FormField.stories.js
@@ -3,10 +3,16 @@ import {
 } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import FormField from '../src/components/FormField.vue';
-import { CUSTOM_INPUT_TYPES, VALID_INPUT_TYPES, VALID_LABEL_STYLES } from '../src/constants';
+import {
+    CUSTOM_INPUT_TYPES,
+    DEFAULT_FIELD_SIZE,
+    VALID_INPUT_TYPES,
+    VALID_FIELD_SIZES,
+    VALID_LABEL_STYLES
+} from '../src/constants';
 
 export default {
-    title: 'Components/Atoms',
+    title: 'Components/Atoms/f-form-field',
     decorators: [withKnobs, withA11y]
 };
 
@@ -18,6 +24,12 @@ export const FormFieldComponent = () => ({
         },
         labelText: {
             default: text('Label Text', 'First name')
+        },
+        fieldSize: {
+            default: select('Field Size', VALID_FIELD_SIZES, DEFAULT_FIELD_SIZE)
+        },
+        isDisabled: {
+            default: select('isDisabled', [null, 'disabled'])
         },
         inputType: {
             default: select('Input Type', VALID_INPUT_TYPES.concat(CUSTOM_INPUT_TYPES))
@@ -39,6 +51,15 @@ export const FormFieldComponent = () => ({
         },
         hasInputDescription: {
             default: boolean('hasExtraLabelText', false)
+        },
+        labelDetails: {
+            default: text('Label Details', '')
+        },
+        assistiveText: {
+            default: text('Assistive Text', '')
+        },
+        placeholder: {
+            default: text('Placeholder', '')
         }
     },
     parameters: {
@@ -50,12 +71,17 @@ export const FormFieldComponent = () => ({
             :label-text="labelText"
             :input-type="inputType"
             :label-style="labelStyle"
+            :field-size="fieldSize"
             :has-error="hasError"
             :dropdown-options="dropdownOptions"
             :is-grouped="isGrouped"
+            :disabled="isDisabled"
             :cols="30"
+            :placeholder="placeholder"
             :rows="7"
             :maxlength="200"
+            :labelDetails="labelDetails"
+            :assistiveText="assistiveText"
             :has-input-description="hasInputDescription">
                 Here is a bit more text to show
         </form-field>`