+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -409,8 +409,7 @@ export default {
         * `state.AuthToken`, then retrieve the phone number from customer api
         * This can happen for newly created guest */
         shouldLoadCustomer () {
-            return this.isLoggedIn &&
-                !this.customer.mobileNumber;
+            return this.isLoggedIn && !this.customer.mobileNumber;
         },
 
         shouldShowCheckoutForm () {