+++ packages/components/organisms/f-content-cards/src/components/cardTemplates/TermsAndConditionsCard.vue
@@ -1,3 +1,4 @@
+<!-- DEPRECATED - we don't need this on the frontend anymore -->
 <template>
     <div
         :data-test-id="dataTestId('shell')"