+++ CHANGELOG.md
@@ -4,6 +4,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v3.50.0
+------------------------------
+*July 22, 2021*
+
+### Added
+- Names for unnamed Circle CI steps.
+
+
+v3.49.0
+------------------------------
+*July 20, 2021*
+
+### Updated
+- Fozzie version to v5.0.0-beta.10 which has the latest pie design token variables.
+
+
+v3.48.0
+------------------------------
+*July 19, 2021*
+
+### Added
+- Test level overview documentation to Storybook.
+
+
 v3.47.0
 ------------------------------
 *July 19, 2021*