+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -4,6 +4,44 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.20.0
+------------------------------
+*July 26, 2021*
+
+### Added
+- `labelDetails` and `assistiveText`.
+- Styling to placeholder.
+
+
+v1.19.0
+------------------------------
+*July 22, 2021*
+
+### Added
+- Leading and Trailing icon slots.
+- Tests to cover changes.
+
+
+v1.18.0
+------------------------------
+*July 22, 2021*
+
+### Changed
+- Made text area data test id dynamic based on name attribute
+
+
+v1.17.0
+------------------------------
+*July 19, 2021*
+
+### Changed
+- Updates to match PIE designs.
+- Disabled label style.
+
+### Added
+- `fieldSize` prop.
+
+
 v1.16.1
 ------------------------------
 *June 30, 2021*