+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,36 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.165.0
+------------------------------
+*July 26, 2021*
+
+### Added
+- `errorCode` property for failed UpdateCheckout is passed to the log
+
+v0.164.0
+------------------------------
+*July 22, 2021*
+
+### Added
+- Added `x-je-feature` to header config in `services/orderPlacementApi`
+
+
+v0.163.0
+------------------------------
+*July 17, 2021*
+
+### Added
+- Extracted APIs from Checkout Store so they are accessible to Contract Tests
+
+
+Latest (to be added to next version)
+------------------------------
+*July 15, 2021*
+
+### Changed
+- Updated version of `f-button`.
+
 
 Latest (to be added to next version)
 ------------------------------