+++ packages/components/organisms/f-loyalty/CHANGELOG.md
@@ -4,6 +4,24 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+Latest (to be added to next release)
+------------------------------
+*July 23, 2021*
+
+### Removed
+- Unused f-button and f-card component dev dependencies
+
+
+v0.2.0
+------------------------------
+*July 20, 2021*
+
+### Added
+- Vuex module for storing data for the loyalty system
+- Registration for the vuex module
+- Base config for globalisation mixin
+
+
 v0.1.0
 ------------------------------
 *June 2, 2021*