+++ packages/services/f-analytics/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.4.0
+------------------------------
+*July 26, 2021*
+
+### Added
+- Route to Feature name constant mapper.
+
 
 v0.3.0
 ------------------------------
@@ -23,6 +30,7 @@ v0.2.0
 v0.1.0
 ------------------------------
 *June 17, 2021*
+
 
 ### Added
 - Initial scaffolding for new component